wayland: port to 1.0 protocol.
Previously some of the functions that this code relied upon were exported as
symbols from the wayland-client .so. However those are now autogenerated
instead and are thus included as static inlines in the header file. Therefore
we must recreated the desired functions using the function pointers found in
the vtable.

Also following the removal of the globals hash from the client code it is
necessary to setup a registry with a listener on it to receive the global
objects.

Signed-off-by: Rob Bradford <rob@linux.intel.com>