Use the wavefront-scan instead of raster-scan during VME prediction on Haswell
When adding the MV prediction for AVC encoding, the GPU thread of current
macroblock needs to wait for the completion of neighbour macroblocks(left/top/up-right).
And the raster-scan will degrade the performance of AVC encoding. So the
wavefront-scan is used to spanw GPU thread, which can avoid the degrading of
the encoding performance.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>