H264_Encoding: Parse the packed header data from user to fix the hacked code of HW skip bytes
When the packed header data from user is inserted into the coded clip, it uses
the hacked code to check the number of HW skip emulation bytes. This is wrong.
So fix it.
Of course if the packed header data is generated by the driver, it is
unnecessary to check it and it can still use the pre-defined number of HW
skip bytes.

V1->V2: Based on Gwenole's comment more nal_unit_type is added.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
(cherry picked from commit cd518563f239cb8523c58010a695098465a4f04e)