Fix scaling of NV12 surfaces when no output_region is set
The code block filling destination surface target rectangle is too early
and incorrectly fills input surface width & height as final target width
and height. This results in improper scaling of the source surface.

Move the destination code block further below the point when actual
target surface is referenced and its width and height are known and use
them to fill the destination rectangle.

Signed-off-by: Antti Seppälä <a.seppala@gmail.com>