gen8_mfd: free surface used by JPEG decode
This is a follow up to previous commit:

commit 8dda0650b55c467200c34ada127b5d9b54edea5c
Author: Zhao Yakui <yakui.zhao@intel.com>
Date:   Tue May 5 17:56:38 2015 +0800

    Decode: Free the allocated internal surface used by JPEG to avoid memory leak

For GEN8, this code path is normally disabled by default.
However, one can enable this code path by defining JPEG_WA
during compilation.  In that case, we would enable the
same memory leak.  Thus, apply the same fix in gen8_mfd to
fix it.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
Signed-off-by: Sean V Kelley <seanvk@posteo.de>
Tested-by: Sean V Kelley <seanvk@posteo.de>