Wrapper the DriverContextP of backend driver
The default directory of backend driver is used.(LIBVA_DRIVERS_PATH).
Only when the backend driver exists, the wrapper is initialized and added.
Otherwise it won't be initialized.

And the option of "enable-wrapper" is added, which is used to determine whether
the wrapper of backend driver is supported.

Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Signed-off-by: Sean V Kelley <seanvk@posteo.de>
(cherry picked from commit 3bbaff8b4180d2510c8c77d9e4d19bf84853f31a)