test: add googletest release-1.8.0 source
We will use the Google Test Framework for testing.

Google Test is not meant to be packaged and used as a system
dependency like other libraries.  Google Test recommends that
it be custom compiled for each project that uses it.  Thus,
we'll keep our own copy of it here and compile it as needed.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>
Reviewed-by: Sean V Kelley <seanvk@posteo.de>
(cherry picked from commit 1746adbb9adabd9fbee417ae3db67da599cc84e9)