i965_drv: fix building with --enable-static
Avoids compiling some unnecessary code in i965_Terminate when
building without HAVE_HYBRID_CODEC.

In particular, this removes the invocation to dlclose(), which
means the driver can be compiled without a dependency on libdl.
Previously, the generated libi965_drv_video.a could not be used
statically because it referenced dlclose(). The corresponding
dlopen() is already guarded behind the HAVE_HYBRID_CODEC flag.

Fixes #27

Signed-off-by: Aman Gupta <aman@tmm1.net>