gen9_avc_encoder: brc_curbe_size is not static anymore
Static memory allocation in general is the allocation of memory at compile
time before the associated program is executed.

And this is the case when compiling with clang: it tries to optimize the
allocation of static memory in compilation time, thus commit f896ca2a breaks
the compilation, since the structure can only be defined at execution time.

Fix #184

Signed-off-by: Víctor Manuel Jáquez Leal <vjaquez@igalia.com>