genX_mfd: fix hardware programming for bitplanes coded in raw mode
Bitplanes that are coded in raw mode are signalled through the
raw_coding flags in the VAPictureParameterBufferVC1 structure. The
absence of a bitplane does not necessarily imply that this bit is
encoded in raw mode. A progressive P picture in advanced profile, for
example, does not encode the SKIPMB bitplane nor does it encode this
bit in raw mode.

Signed-off-by: Jerome Borsboom <jerome.borsboom@carpalis.nl>