genX_mfd: clean up the use of profile bitstream element
Only gen6 needs mapping from input range to the hardware range for
the profile bitstream element. For gen7 and up there is no need for a
mapping as the hardware uses only one bit indicating simple/main profile
or advanced profile. Therefore, we can remove the mapping for gen7 and
clean up the code a bit.

Signed-off-by: Jerome Borsboom <jerome.borsboom@carpalis.nl>