test: fix use of uninitialized variable
Initialize VAImage variable at the point of declaration.

Since i965_DeriveImage is declared as extern in the tests,
a static scanner appears to be unable to detect that VAImage is
actually initialized by the i965_DeriveImage implementation
in the driver.  This causes it to flag it as "use of an
uninitialized variable" along the success path.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>