Fix packed headers attribute test in vaCreateConfig()
If packed headers are supported, then the provided value must be some
(possibly empty) subset of the supported headers.

This fixes config creation with an empty packed header set, which the
user may pass if they don't want to provide any packed headers.  This
pattern is used in at least libavcodec, where VP9 encoding is broken
prior to this change.

Since there is intent to deprecate this behaviour in future, also add
warnings to this case and the other possible failure modes here.

Fixes #362.

Signed-off-by: Mark Thompson <sw@jkqxz.net>