Remove implicit truncations from int to bit-field
All the members in the structures scoreboard1 and scoreboard2 are
signed int of 4 bit long (this means they only admit values from
-8 to 7).

Along the code, these structures are assigned with values bigger
than 7 which are implicitly truncated to negative numbers. For
example, 0xF -> -1 or 0xE -> -2

It is better to explictly show the assigned values than rely on
implicit convertions. This patch aims to that direction.