Remove implicit truncation from unsigned int to bit field
All the members of struct i965_sampler_8x8_coefficient are signed
int of 8 bit long (ranging from -128 to 127).

But there is an assignation of 1U<<7 which is implicitly truncated
to -128.

To make explicit the casting, without losing the expresiveness of
the bitwise operation, this patch adds a macro which uses the
anonymous union technique.