Use f_code to determine max MV length
The motion vector length is constrained by the level but not set by it -
we should instead use the f_code values to set the max MV length.

This fixes encoding when the f_code values are not set to the highest
allowed for the level, and also adds support for intermediate levels such
as High 1440 which were not previously handled here.

Signed-off-by: Mark Thompson <sw@jkqxz.net>