i965_encoder: Speed up i965_MapBuffer for JPEG encoding somewhat.
Searching for the EOI marker byte-by-byte turned out to notable in
profiles when encoding large amounts of JPEGs (~15% of a core at 480 fps
of 1080p images). Using memmem() will typically give us an AVX-optimized
version of at least finding the 0xFF character, which is much more
efficient. It seems to speed up this part by about 3–4x in practice,
taking it largely off the profiles.

Signed-off-by: Steinar H. Gunderson <steinar+nageru@gunderson.no>