meson: fix warning about unexpected return code checking for run_command
Recent versions of Meson warn you that the default is to not check the
return code, which is a bad default and may eventually change. To
suppress this warning, an explicit `check: ` value must be set.

Considering the code in play here:
- check if git exists
- if so, always assume this is running from a git checkout
- embed either '()' or '(git describe version)'

it seems likely the intention is indeed to have it be `check: false`,
but there's some missing error checking here to ensure it.

Check the returncode. If git fails, it is surely because there is no git
repository and the build is being run from a tarball. In that case,
behave as though git wasn't found in the first place, and use the
fallback value.

Suppressing the warning means bumping the minimum version of Meson. This
can be safely done since both the previous and new minimums are quite
old, and libva already depends on a much newer version.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>